pyproject.toml
[project]
name = "tg-dida"
version = "0.1.0"
description = "Telegram Bot for Dida365 task management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
tg-dida = "main:main"
